requirements.txt
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
typed-ast==1.3.5
